fix: correct split slice board
