+++ src/interfaces/user.ts
@@ -1,0 +1,5 @@
+export interface IUser {
+  id: string;
+  name: string;
+  login: string;
+}
