+++ src/store/reducers/actions/board.ts
@@ -4,6 +4,7 @@ import { AxiosError } from 'axios';
 import { IBoard, ResponceBoard, ResponceBoard as ResponseBoard } from 'interfaces/boards';
 import { axiosErrorHandler } from 'utils/helpers';
 import { API_ENDPOINTS } from 'utils/variables';
+import { IRequestForBoard } from 'interfaces/boards';
 
 type PayloadActionError = PayloadAction<string>;
 
@@ -16,6 +17,19 @@ export const getAllBoards = createAsyncThunk<Array<IBoard>, void, { rejectValue:
     } catch (error) {
       const err = error as AxiosError;
       return rejectWithValue('dddd');
+    }
+  }
+);
+
+export const createBoard = createAsyncThunk(
+  'boards/create',
+  async (dataBoardCreator: IRequestForBoard, { rejectWithValue }) => {
+    try {
+      const response = await axiosPrivate.post(API_ENDPOINTS.BOARDS, dataBoardCreator);
+      return response.data;
+    } catch (error) {
+      const err = error as AxiosError;
+      return rejectWithValue(axiosErrorHandler(err));
     }
   }
 );
