+++ src/utils/helpers.ts
@@ -25,5 +25,6 @@ export const parseJwt = (token: string) => {
       })
       .join('')
   );
+  console.log(JSON.parse(jsonPayload));
   return JSON.parse(jsonPayload);
 };
